--- back_end/capstone/build.gradle
@@ -1,48 +1,0 @@
-plugins {
-	id 'java'
-	id 'org.springframework.boot' version '3.4.4'
-	id 'io.spring.dependency-management' version '1.1.7'
-}
-
-group = 'com.example'
-version = '0.0.1-SNAPSHOT'
-
-java {
-	toolchain {
-		languageVersion = JavaLanguageVersion.of(21)
-	}
-}
-
-repositories {
-	mavenCentral()
-}
-
-dependencies {
-	implementation 'org.springframework.boot:spring-boot-starter-web'
-	testImplementation 'org.springframework.boot:spring-boot-starter-test'
-	testRuntimeOnly 'org.junit.platform:junit-platform-launcher'
-	
-	developmentOnly 'org.springframework.boot:spring-boot-devtools'
-	
-	compileOnly 'org.projectlombok:lombok' 
-    annotationProcessor 'org.projectlombok:lombok'
-    
-    implementation 'org.springframework.boot:spring-boot-starter-data-mongodb'
-    
-    implementation 'org.springframework.boot:spring-boot-starter-data-jpa'
-    
-    implementation 'org.springframework.boot:spring-boot-starter-security'
-    implementation 'org.thymeleaf.extras:thymeleaf-extras-springsecurity6'
-    implementation 'io.jsonwebtoken:jjwt-api:0.11.5'
-   implementation 'io.jsonwebtoken:jjwt-impl:0.11.5'
-   implementation 'io.jsonwebtoken:jjwt-jackson:0.11.5'
-    
-    testImplementation 'org.junit.jupiter:junit-jupiter' 
-    testRuntimeOnly 'org.junit.platform:junit-platform-launcher'
-    
-    implementation 'org.springframework.boot:spring-boot-starter-validation'
-}
-
-tasks.named('test') {
-	useJUnitPlatform()
-}
