--- back_end/capstone/src/main/java/com/defuture/stockapp/CapstoneApplication.java
@@ -1,14 +1,0 @@
-package com.defuture.stockapp;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class CapstoneApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(CapstoneApplication.class, args);
-	}
-
-}
-
