--- back_end/capstone/src/main/java/com/defuture/stockapp/SecurityConfig.java
@@ -1,35 +1,0 @@
-package com.defuture.stockapp;
-
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.security.config.annotation.web.builders.HttpSecurity;
-import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
-import org.springframework.security.web.SecurityFilterChain;
-import org.springframework.security.web.authentication.UsernamePasswordAuthenticationFilter;
-import org.springframework.security.web.util.matcher.AntPathRequestMatcher;
-
-import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
-import org.springframework.security.crypto.password.PasswordEncoder;
-
-@Configuration
-@EnableWebSecurity
-public class SecurityConfig {
-	
-	@Bean
-    SecurityFilterChain filterChain(HttpSecurity http) throws Exception {
-        http
-        	.csrf(csrf -> csrf.disable())
-            .authorizeHttpRequests((authorizeHttpRequests) -> authorizeHttpRequests
-                .requestMatchers(new AntPathRequestMatcher("/**")).permitAll())
-            .formLogin((formLogin) -> formLogin
-                    .loginPage("/api/users/login")
-                    .defaultSuccessUrl("/"))
-        ;
-        return http.build();
-    }
-	
-	@Bean
-    PasswordEncoder passwordEncoder() {
-        return new BCryptPasswordEncoder();
-    }
-}
