--- back_end/capstone/src/main/java/com/defuture/stockapp/users/UserRepository.java
@@ -1,11 +1,0 @@
-package com.defuture.stockapp.users;
-
-import java.util.Optional;
-
-import org.springframework.data.mongodb.repository.MongoRepository;
-
-public interface UserRepository extends MongoRepository<UserAccount, String>{
-	Optional<UserAccount> findByUserId(String userId);
-	boolean existsByUserId(String userId);
-	boolean existsByEmail(String email);
-}
