--- back_end/capstone/src/main/resources/application.properties
@@ -1,15 +1,0 @@
-spring.application.name=capstone
-
-# JPA
-spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.H2Dialect
-spring.jpa.hibernate.ddl-auto=update
-
-
-spring.data.mongodb.uri=mongodb+srv://jun:{password}@cluster0.jx94q.mongodb.net/user_accounts?retryWrites=true&w=majority&appName=Cluster0
-spring.autoconfigure.exclude=org.springframework.boot.autoconfigure.jdbc.DataSourceAutoConfiguration
-spring.data.mongodb.auto-index-creation=true
-
-securities.api.base-url=https://mockapi.kiwoom.com
-securities.api.token-endpoint=/oauth2/token
-securities.api.account-endpoint=/api/dostk/acnt
-
