--- back_end/capstone/src/test/java/com/defuture/stockapp/CapstoneApplicationTests.java
@@ -1,13 +1,0 @@
-package com.defuture.stockapp;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class CapstoneApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
