+++ back_end/settings.gradle
@@ -1,0 +1,1 @@
+rootProject.name = 'back_end'
