+++ back_end/src/main/java/com/defuture/stockapp/AppConfig.java
@@ -1,0 +1,14 @@
+package com.defuture.stockapp;
+
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.web.client.RestTemplate;
+
+@Configuration
+public class AppConfig {
+
+    @Bean
+    public RestTemplate restTemplate() {
+        return new RestTemplate();
+    }
+}
