+++ back_end/src/main/java/com/defuture/stockapp/BackEndApplication.java
@@ -1,0 +1,13 @@
+package com.defuture.stockapp;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class BackEndApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(BackEndApplication.class, args);
+	}
+
+}
