+++ back_end/src/main/java/com/defuture/stockapp/jwt/AuthenticationRequestDTO.java
@@ -1,0 +1,30 @@
+package com.defuture.stockapp.jwt;
+
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+@Getter
+@Setter
+@NoArgsConstructor
+public class AuthenticationRequestDTO {
+	 // 로그인 시 필요한 정보: username과 password
+    private String username;
+    private String password;
+    
+    public String getUsername() {
+        return username;
+    }
+    
+    public void setUsername(String username) {
+        this.username = username;
+    }
+
+    public String getPassword() {
+        return password;
+    }
+
+    public void setPassword(String password) {
+        this.password = password;
+    }
+}
