+++ back_end/src/test/java/com/defuture/stockapp/BackEndApplicationTests.java
@@ -1,0 +1,13 @@
+package com.defuture.stockapp;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class BackEndApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
